+++ CHANGES.md
@@ -1,7 +1,7 @@
 ## Changes in version 0.4.5 (in development)
 
-* Users can now download time-series data as a Zip file
-  containing a JSON file for each time-series graph. (#20)
+* Users can now download time-series data as a Zip archive
+  containing a GeoJSON file for each time-series graph. (#20)
 
 
 ## Changes in version 0.4.4